+++ src/methods/cycombine_one_control_to_mid/script.R
@@ -6,7 +6,7 @@ par <- list(
     input = "resources_test/task_cyto_batch_integration/leomazzi_cyto_spleen_subset/unintegrated_censored.h5ad",
     output = "resources_test/task_cyto_batch_integration/leomazzi_cyto_spleen_subset/output.h5ad"
 )
-meta <- list(name = "cycombine_all_controls")
+meta <- list(name = "cycombine_one_control")
 ## VIASH END
 
 cat("Reading input files\n")
@@ -23,8 +23,8 @@ adata_to_correct <- input_adata[, markers_to_correct]
 df_to_correct <- as.data.frame(
     adata_to_correct$layers[["preprocessed"]]
 )
-df_to_correct$batch <- adata_to_correct$obs$batch
-df_to_correct$sample <- adata_to_correct$obs$sample
+df_to_correct$batch <- as.factor(adata_to_correct$obs$batch)
+df_to_correct$sample <- as.factor(adata_to_correct$obs$sample)
 
 # add an "anchor" column which specify which samples are the technical replicates
 # this is a bit weird as the anchor information should be a column name or character vector
@@ -38,10 +38,11 @@ df_to_correct$sample <- adata_to_correct$obs$sample
 # the as.character function is needed as otherwise we will get NAs for those controls..
 
 df_to_correct$anchor <- ifelse(
-    adata_to_correct$obs$is_control == 0,
-    as.character(adata_to_correct$obs$sample),
-    paste0("control_", adata_to_correct$obs$is_control)
+    adata_to_correct$obs$is_control == 1,
+    paste0("control_", adata_to_correct$obs$is_control),
+    as.character(adata_to_correct$obs$sample)
 )
+df_to_correct$anchor <- as.factor(df_to_correct$anchor)
 
 lineage_markers <- as.vector(input_adata$var_names[
     input_adata$var$marker_type == "lineage"
@@ -52,8 +53,6 @@ cat("Run cyCombine\n")
 
 # use the default parameters in normalize
 # do the normalisation on all markers to be corrected
-# use z-score normalisation because we are merging batches
-# from a single study.
 df_to_correct_norm <- cyCombine::normalize(
     df = df_to_correct,
     markers = markers_to_correct,